+++ collection.h
@@ -17,7 +17,7 @@ public:
     collection(std::string filename, const catalog &catalog);
 
     card at(uint8_t loc) const;
-    uint8_t count();
+    uint64_t count();
 private:
     const catalog &catalog_;
     std::vector<std::string> cards_;
